Fixed data type

+++ src/application/dao/GameDAO.java
@@ -12,7 +12,7 @@ public interface GameDAO {
 	List<Game> getGamesByUser(int userId) throws SQLException;
 	int addGenre(String genre) throws SQLException;
 	void linkGameGenre(int gameId, int genreId) throws SQLException;
-	void updateGameRating(String title, int rating) throws SQLException;
+	void updateGameRating(String title, double rating) throws SQLException;
 	void updateReview(String title, String review) throws SQLException;
 	void deleteGame(String title) throws SQLException;
 }
